Add minimal parser implementation (Recursive option)
The task to built an IncludedResources decoder imposes 3 problems:
  - converting nested resources into respective hash representations
  - merging hashes
  - mitigating collisions between nested and plain resources

Proposed implementation starts by segregating singular resources in a
separate array which will be later merged with nested resources
on lower priority.
That effectively eleminates possibility of double inclusion of a nested
and a plain resource with the same name.

Nested resources are decoded and then merged together in a resultive array.
Notice, that in-place decoding and replacement might be more effective.

Known bugs:
 - Decoder can't properly merge resources with differing length